BackEnd/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
python-multipartemail-validator
